Split Session into provider/controller
